Feat: Build Processor Regex added tests
