+++ src/hints/hint_processor/function_like_hint_processor/test_cairo_programs/wrong_custom_hint.cairo
@@ -1,0 +1,5 @@
+func test_wrong_custom_hint() {
+    %{ wrong_custom_hint) wrong ( %}
+    assert 2 = 3;
+    return ();
+}
